+++ src/record/settings.rs
@@ -242,15 +242,9 @@ impl RecordSettings {
 						_ => None,
 					},
 					u64::from_str_radix(color, 16).unwrap_or(Self::default().color),
-					if matches.is_present("no-borders") {
-						None
-					} else if !padding.is_zero() {
-						Some(1)
-					} else {
-						Some(parser.parse(
-							"border",
-							Self::default().border.unwrap_or_default(),
-						))
+					match parser.parse("border", 0) {
+						border if border > 0 => Some(border),
+						_ => None,
 					},
 					padding,
 					RecordTime::from_parser(&parser),
@@ -314,7 +308,6 @@ mod tests {
 			.arg(Arg::with_name("root").long("root"))
 			.arg(Arg::with_name("focus").long("focus"))
 			.arg(Arg::with_name("with-alpha").long("with-alpha"))
-			.arg(Arg::with_name("no-borders").long("no-borders"))
 			.arg(Arg::with_name("no-keys").long("no-keys"))
 			.get_matches_from(vec![
 				"test",