+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update contribution guidelines about changes in command line arguments
 - Update the optimization level of the development profile
 - Build with debug profile while generating shell completions
+- Use `--border` argument for setting the border width
 
 ### Removed
 - Remove update_release action from CD workflow